Fixed functions
