+++ app/routes.py
@@ -1,4 +1,4 @@
-from app import app, tasks, getInfo
+from app import app, tasks, getInfo, cronTasks, queries
 from flask import render_template, redirect, request
 from apscheduler.schedulers.background import BackgroundScheduler
 import time
@@ -7,6 +7,7 @@ import atexit
 
 scheduler = BackgroundScheduler()
 meterRead = app.config['METERREAD']
+
 
 @app.before_first_request
 def startBackGroundJob():
@@ -17,20 +18,20 @@ def startBackGroundJob():
         max_instances=1
     )
     scheduler.add_job(
-        tasks.dailyTasks,
+        cronTasks.dailyTasks,
         trigger='cron',
         hour='23',
         minute='59'
     )
     scheduler.add_job(
-        tasks.weeklyTasks,
+        cronTasks.weeklyTasks,
         trigger='cron',
         day_of_week='sun',
         hour='0',
         minute='0'
     )
     scheduler.add_job(
-        tasks.monthlyTasks,
+        cronTasks.monthlyTasks,
         trigger='cron',
         day='%(s)s' % {'s': meterRead},
         hour='0',
@@ -46,20 +47,20 @@ def startBackGroundJob():
 
 @app.route('/')
 def index():
-    print(tasks.tskQryToList(tasks.qryCurrent()))
+    print(tasks.tskQryToList(queries.qryCurrent()))
     return render_template(
         'index.html',
-        currentStatus=tasks.qryCurrent(),
-        voltageStats=tasks.qryVoltage(),
-        killawattStats=tasks.qryKillawatt(),
-        dayKwhTotal=tasks.qryDayKwhTotal(),
-        kwh7dTotal=tasks.qryKwh7dTotal(),
-        kwhPrevWk=tasks.qryKwhPrevWk(),
-        kwhPrevMn=tasks.qryKwhPrevMn(),
+        currentStatus=queries.qryCurrent(),
+        voltageStats=queries.qryVoltage(),
+        killawattStats=queries.qryKillawatt(),
+        dayKwhTotal=queries.qryDayKwhTotal(),
+        kwh7dTotal=queries.qryKwh7dTotal(),
+        kwhPrevWk=queries.qryKwhPrevWk(),
+        kwhPrevMn=queries.qryKwhPrevMn(),
         kwhCost=tasks.tskCalculateCost(),
-        peakKwhDayMn=tasks.qryPeakKwhDayMn(),
-        lowKwhDayMn=tasks.qryLowKwhDayMn(),
-        avgKwhDayMn=tasks.qryAvgKwhDayMn(),
+        peakKwhDayMn=queries.qryPeakKwhDayMn(),
+        lowKwhDayMn=queries.qryLowKwhDayMn(),
+        avgKwhDayMn=queries.qryAvgKwhDayMn(),
         bills=tasks.tskGetBills(),
         pws=app.config['PWS']
     )
@@ -67,8 +68,8 @@ def index():
 
 @app.route('/rtkw')
 def rtkw():
-    currentStatus = tasks.qryCurrent()
-    value = str(currentStatus[0][1])
+    currentStatus = queries.qryCurrent()
+    value = str(currentStatus[0][2])
     return value
 
 
@@ -80,9 +81,11 @@ def about():
         dbver=app.config['DBVER']
     )
 
+
 @app.route('/bills')
 def bills():
     return render_template('about.html')
+
 
 @app.route('/billData')
 def billData():
@@ -95,10 +98,11 @@ def billData():
         billData=tasks.tskGetBillingData(bid)
     )
 
+
 @app.route('/runtasks')
 def runTasks():
-    tasks.dailyTasks()
-    tasks.monthlyTasks()
+    cronTasks.dailyTasks()
+    cronTasks.monthlyTasks()
     return redirect('/')
